import optimization fix: balance is now refreshed on restart of the application. To do so, the getUser() method has been moved from ConnexionActivity to MainActivity.
