+++ app/src/main/java/com/example/mycomifclient/connexion/FirstConnexionActivity.kt
@@ -2,7 +2,6 @@ package com.example.mycomifclient.connexion
 
 import android.content.Context
 import android.os.Bundle
-import android.view.View
 import android.widget.Button
 import android.widget.ImageButton
 import androidx.appcompat.app.AlertDialog
